feat: Add a staging step before release
Previously once you ended the changelog editing step,
the release would just go through immediately.

I think this violated the principle of least surprise
as for something as big as a new release, I think
it's probably good to have a step in which you first
check that the release that is about to go out is legit.